Warning fixes for implicite casts that vs2012 complains about with higher warning levels BUG=302 TESTED=hammer build R=tpsiaki@google.com
Review URL: https://webrtc-codereview.appspot.com/6559004

git-svn-id: http://libyuv.googlecode.com/svn/trunk@948 16f28f9a-4ce2-e073-06de-1de4eb20be90

+++ source/scale_argb.cc
@@ -178,7 +178,7 @@ static void ScaleARGBBilinearDown(int src_width, int src_height,
   xr = (xr >> 16) + 1;  // Right most pixel used.
   int clip_src_width = (((xr - xl) + 1 + 3) & ~3) * 4;  // Width aligned to 4.
   src_argb += xl * 4;
-  x -= (xl << 16);
+  x -= static_cast<int>(xl << 16);
   void (*InterpolateRow)(uint8* dst_argb, const uint8* src_argb,
       ptrdiff_t src_stride, int dst_width, int source_y_fraction) =
       InterpolateRow_C;
